chore(serverless-contracts): move bench files closer to source files
