--- packages/serverless-contracts/src/__benches__/contracts/index.ts
@@ -1,2 +1,0 @@
-export * from './apiGateway';
-export * from './eventBridge';
