+++ packages/serverless-contracts/src/contracts/apiGateway/__mocks__/httpApiGatewayContract.bench.ts
