+++ packages/serverless-contracts/src/contracts/eventBridge/__benches__/basicEventBridgeHandler.bench.ts
