+++ packages/serverless-contracts/src/contracts/eventBridge/__benches__/bigEventBridgeHandler.bench.ts
