+++ packages/serverless-contracts/src/contracts/eventBridge/__benches__/index.ts
